# Improved Serializer memory usage