--- a/Source/MQTTnet/Serializer/IMqttPacketSerializer.cs
+++ b/Source/MQTTnet/Serializer/IMqttPacketSerializer.cs
@@ -8,7 +8,7 @@ namespace MQTTnet.Serializer
     {
         MqttProtocolVersion ProtocolVersion { get; set; }
 
-        ArraySegment<byte> Serialize(MqttBasePacket mqttPacket);
+        byte[] Serialize(MqttBasePacket mqttPacket);
 
         MqttBasePacket Deserialize(ReceivedMqttPacket receivedMqttPacket);
     }