--- a/Source/MQTTnet/Serializer/MqttPacketReader.cs
+++ b/Source/MQTTnet/Serializer/MqttPacketReader.cs
@@ -1,4 +1,6 @@
-﻿using System.Threading;
+﻿using System;
+using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using MQTTnet.Channel;
 using MQTTnet.Exceptions;
@@ -43,6 +45,44 @@ namespace MQTTnet.Serializer
             return new MqttFixedHeader(buffer[0], bodyLength);
         }
 
+        public static byte ReadByte(this ref ReadOnlySpan<byte> stream)
+        {
+            var result = stream[0];
+            stream = stream.Slice(1);
+            return result;
+        }
+
+        public static ushort ReadUInt16(this ref ReadOnlySpan<byte> stream)
+        {
+            var result = System.Buffers.Binary.BinaryPrimitives.ReadUInt16BigEndian(stream);
+            stream = stream.Slice(2);
+            return result;
+        }
+
+        public static string ReadStringWithLengthPrefix(this ref ReadOnlySpan<byte> stream)
+        {
+            var buffer = stream.ReadWithLengthPrefix();
+            if (buffer.Length == 0)
+            {
+                return string.Empty;
+            }
+
+            return Encoding.UTF8.GetString(buffer, 0, buffer.Length);
+        }
+
+        public static byte[] ReadWithLengthPrefix(this ref ReadOnlySpan<byte> stream)
+        {
+            var length = stream.ReadUInt16();
+            if (length == 0)
+            {
+                return new byte[0];
+            }
+
+            var result = stream.Slice(0, length).ToArray();
+            stream = stream.Slice(length);
+            return result;
+        }
+        
         private static int ReadBodyLength(IMqttChannel channel, byte initialEncodedByte, CancellationToken cancellationToken)
         {
             // Alorithm taken from https://docs.oasis-open.org/mqtt/mqtt/v3.1.1/errata01/os/mqtt-v3.1.1-errata01-os-complete.html.