--- a/Tests/MQTTnet.Benchmarks/MessageProcessingBenchmark.cs
+++ b/Tests/MQTTnet.Benchmarks/MessageProcessingBenchmark.cs
@@ -9,6 +9,7 @@ namespace MQTTnet.Benchmarks
 {
     [ClrJob]
     [RPlotExporter, RankColumn]
+    [MemoryDiagnoser]
     public class MessageProcessingBenchmark
     {
         private IMqttServer _mqttServer;