… Tests/MQTTnet.Core.Tests/MqttPacketSerializerTests.cs