before:

|                      Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
| --------------------------- |---------:|----------:|----------:|----------:|----------:|
|    Serialize_10000_Messages | 2.496 ms | 0.0463 ms | 0.0410 ms |  781.2500 |   3.13 MB |
|  Deserialize_10000_Messages | 7.791 ms | 0.0842 ms | 0.0746 ms | 3765.6250 |  15.11 MB |

after:

|                      Method |      Mean |     Error |    StdDev |     Gen 0 |  Allocated |
| --------------------------- |---------:|----------:|----------:|----------:|-----------:|
|    Serialize_10000_Messages |     2.830 ms | 0.0558 ms | 0.0597 ms |  152.3438 |  625.01 KB |
|  Deserialize_10000_Messages |      5.313 ms | 0.0790 ms | 0.0739 ms | 1218.7500 | 5000.15 KB |

we can argue about serializer performance as it is 10% slower now but I guess the memory pooling should be worth it especially for bigger messages. anyway let me know what u think about this.

there is still some potential in string encoding MqttPacketWriter.cs line 33 - 35 but that will have to wait until the package is final.

preview with that  change:

|                      Method |     Mean |     Error |    StdDev |     Gen 0 |  Allocated |
| --------------------------- |---------:|----------:|----------:|----------:|-----------:|
|    Serialize_10000_Messages | 2.613 ms | 0.0507 ms | 0.0623 ms |   74.2188 |   312.5 KB |
|  Deserialize_10000_Messages | 5.488 ms | 0.1090 ms | 0.1455 ms | 1218.7500 | 5000.15 KB |

the header deserialization can also be improved if we use new connection abstractions (see issue #250 )