@JanEggers 2018-06-06
updated performance improvements after rebase:

|                      Method | Iterations |     Mean |     Error |    StdDev |     Gen 0 |  Allocated |
| --------------------------- |----------- |---------:|----------:|----------:|----------:|-----------:|
|    Serialize_10000_Messages |      10000 | 2.768 ms | 0.0609 ms | 0.0770 ms |  152.3438 |  625.01 KB |
|  Deserialize_10000_Messages |      10000 | 4.560 ms | 0.0901 ms | 0.1578 ms | 1218.7500 | 5000.15 KB |