@JanEggers 2018-06-16
@chkr1011 can you please drop a comment if you cnsider accepting this. and if not close it i have done 3 rebases already and will do a final one if you intend to merge