@chkr1011 2018-06-16
I already pushed a different strategy. But I will compare this results with the results of my change. If this one is faster I will consider using this PR. Otherwise I will keep my optimization.