@JanEggers 2018-06-16
I rebased again and this is the result:

without my changes:

|                     Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
|--------------------------- |---------:|----------:|----------:|----------:|----------:|
|   Serialize_10000_Messages | 2.835 ms | 0.0386 ms | 0.0323 ms |  781.2500 |   3.13 MB |
| Deserialize_10000_Messages | 8.068 ms | 0.0499 ms | 0.0442 ms | 3671.8750 |  14.73 MB |

with my changes:

|                     Method | Iterations |     Mean |     Error |    StdDev |     Gen 0 |  Allocated |
| --------------------------- |----------- |---------:|----------:|----------:|----------:|-----------:|
|   Serialize_10000_Messages |      10000 | 2.920 ms | 0.0562 ms | 0.0730 ms |  152.3438 |  625.01 KB |
| Deserialize_10000_Messages |      10000 | 5.344 ms | 0.0431 ms | 0.0382 ms | 1195.3125 | 4922.02 KB |

so serializing is a little slower but deserialing is nearly twice as fast, but the main advantage is that memory usage is much lower, this will come more into play with bigger messanges and larger topic strings