@chkr1011 2018-06-16
Good work! That's a lot better than mine. But I still have to check what needs to be done regarding the copy of the code from the other project. I will read the license and see if we can just copy it or if we have to include some more details and copyright texts etc.

Maybe I am also able to improve my solution 😄