@JanEggers 2018-06-16
that might help:

https://tldrlegal.com/license/apache-license-2.0-(apache-2.0)