@chkr1011 2018-06-17
Interesting reference. Thanks 👍 
I accepted the challenge to get a better performance without include code from other projects 😄. I am on a good way. I first optimized the benchmark because it was not fair. The the benchmark itself created a lot of memory consuming buffers. I fixed this and already migrated the "indea" behind the SingalR code and there are my results:

|                     Method |     Mean |     Error |    StdDev |    Gen 0 |  Allocated |
|--------------------------- |---------:|----------:|----------:|---------:|-----------:|
|   Serialize_10000_Messages | 1.147 ms | 0.0053 ms | 0.0049 ms | 101.5625 |  625.01 KB |
| Deserialize_10000_Messages | 2.252 ms | 0.0119 ms | 0.0105 ms | 964.8438 | 5937.55 KB |

I believe that the different execution time is caused by different machines. The interesting thing is that the memory for serializing is now exactly the same.

I am still working on the deserialize code.