@chkr1011 2018-06-17
I updated the values.

The buffer is part of the ThreadStatic _MqttPacketWriter_. If the code is really on a different thread it will have a dedicated instance.