@JanEggers 2018-06-17
that is exactly my point so if you have a large message build up a buffer. then async send it and land on a differnt thread then the clear will have no effect since the buffer is on the original thread