@chkr1011 2018-06-17
OK now I got it. Mhhh. AsyncLocal is not available for .net452. I will try to find another solution.