@JanEggers 2018-06-17
anyway I will close this pr as your results are now very close to my my results with system.memory and System.Buffers